requirements.txt
Django>=4.0,<6.0
Pillow
requests
tensorflow==2.7.0
numpy<1.24.0
keras==2.7.0
protobuf==3.20.3
tensorflow-addons==0.15.0
